#define CUBE macro

+++ main.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 
 
-int process_data(int data);
+float process_data(int data);
 
 int load_data()
 {
@@ -16,7 +16,7 @@ void save_data(int data)
 int main()
 {
     int data = load_data();
-    int result = process_data(data);
+    float result = process_data(data);
     save_data(result);
 
     return 1;
